services/enrich-weather/requirements.txt
Flask
requests
boto3
google-cloud-pubsub
